Refactor UpdateRepository for code reuse
