+++ activity/activity_UpdateRepository.py
@@ -1,19 +1,11 @@
 from ssl import SSLError
 import tempfile
-from github import Github
-from github import GithubException
-from provider.utils import pad_msid, settings_environment, unicode_encode
 import provider.lax_provider
+from provider import github_provider
+from provider.github_provider import RetryException
+from provider.utils import pad_msid, settings_environment, unicode_encode
 from provider.storage_provider import storage_context
 from activity.objects import Activity
-
-"""
-activity_UpdateRepository.py activity
-"""
-
-
-class RetryException(RuntimeError):
-    pass
 
 
 class activity_UpdateRepository(Activity):
@@ -76,7 +68,6 @@ class activity_UpdateRepository(Activity):
             return True
 
         try:
-
             xml_file = provider.lax_provider.get_xml_file_name(
                 self.settings,
                 pad_msid(data["article_id"]),
@@ -101,8 +92,11 @@ class activity_UpdateRepository(Activity):
 
                 file_content = storage.get_resource_as_string(resource)
 
-                message = self.update_github(
-                    self.settings.git_repo_path + xml_file, unicode_encode(file_content)
+                message = github_provider.update_github(
+                    self.settings,
+                    self.logger,
+                    self.settings.git_repo_path + xml_file,
+                    unicode_encode(file_content),
                 )
                 self.logger.info(message)
                 self.emit_monitor_event(
@@ -145,61 +139,3 @@ class activity_UpdateRepository(Activity):
                 "Error Updating repository for article. Details: " + str(exception),
             )
             return self.ACTIVITY_PERMANENT_FAILURE
-
-    def update_github(self, repo_file, content):
-
-        github_object = Github(self.settings.github_token)
-        user = github_object.get_user("elifesciences")
-        article_xml_repo = user.get_repo(self.settings.git_repo_name)
-
-        try:
-            xml_file = article_xml_repo.get_contents(repo_file)
-        except GithubException as exception:
-            self.logger.info("GithubException - description: " + str(exception))
-            self.logger.info(
-                "GithubException: file "
-                + repo_file
-                + " may not exist in github yet. We will try to add it in the repo."
-            )
-            try:
-                response = article_xml_repo.create_file(
-                    repo_file, "Creates XML", content
-                )
-            except GithubException as inner_exception:
-                self._retry_or_cancel(inner_exception)
-            return "File " + repo_file + " successfully added. Commit: " + str(response)
-
-        except Exception as exception:
-            self.logger.info(
-                "Exception: file " + repo_file + ". Error: " + str(exception)
-            )
-            raise
-
-        try:
-            # check for changes first
-            if content == xml_file.decoded_content:
-                return "No changes in file " + repo_file
-
-            # there are changes
-            try:
-                response = article_xml_repo.update_file(
-                    repo_file, "Updates xml", content, xml_file.sha
-                )
-            except GithubException as exception:
-                self._retry_or_cancel(exception)
-            return (
-                "File " + repo_file + " successfully updated. Commit: " + str(response)
-            )
-
-        except Exception as exception:
-            self.logger.info(
-                "Exception: file " + repo_file + ". Error: " + str(exception)
-            )
-            raise
-
-    def _retry_or_cancel(self, exception):
-        if exception.status == 409:
-            self.logger.warning("Retrying because of exception: %s" % exception)
-            raise RetryException(str(exception))
-
-        raise exception
